+++ src/carousel/index.ts
@@ -1,0 +1,2 @@
+export type { CarouselState, CarouselProps, CarouselControlData } from './types';
+export { Carousel } from './carousel';
