+++ src/text-button/index.ts
@@ -1,0 +1,8 @@
+export type {
+  TextButtonSize,
+  TextButtonColor,
+  TextButtonProps,
+  TextButtonAsAnchorProps,
+  TextButtonAsButtonProps,
+} from './types';
+export { TextButton } from './text-button';
