+++ src/textarea/index.ts
@@ -1,0 +1,2 @@
+export type { TextareaProps } from './types';
+export { Textarea } from './textarea';
