+++ src/timer/index.ts
@@ -1,0 +1,2 @@
+export type { TimeDistance, TimerProps } from './types';
+export { Timer } from './timer';
