+++ src/toggle/index.ts
@@ -1,0 +1,2 @@
+export type { ToggleProps } from './types';
+export { Toggle } from './toggle';
