+++ src/toggle/toggle.tsx
@@ -1,27 +1,25 @@
+import type { ToggleProps } from './types';
 import { forwardRef } from 'react';
-import { CheckboxInputProps } from '../types';
-import classnames from 'classnames/bind';
+import classNames from 'classnames/bind';
 import styles from './toggle.module.scss';
 
-export type ToggleProps = CheckboxInputProps;
-
-const cx = classnames.bind(styles);
+const cx = classNames.bind(styles);
 
 /**
  * Тогл - элемент для управления полем с двумя состояниями - включено и отключено.
  */
-export const Toggle = forwardRef<HTMLInputElement, ToggleProps>(
-  (
-    {
-      className,
-      style,
-      'data-testid': testId = 'toggle',
+export const Toggle = forwardRef<HTMLInputElement, ToggleProps>(function Toggle(
+  {
+    className,
+    style,
+    'data-testid': testId = 'toggle',
 
-      // input props
-      ...restProps
-    },
-    ref,
-  ) => (
+    // input props
+    ...restProps
+  },
+  ref,
+) {
+  return (
     <label className={cx('root', className)} style={style}>
       <input
         {...restProps}
@@ -32,7 +30,7 @@ export const Toggle = forwardRef<HTMLInputElement, ToggleProps>(
       />
       <span className={cx('switch')} />
     </label>
-  ),
-);
+  );
+});
 
 Toggle.displayName = 'Toggle';
