+++ src/toggle/types.ts
@@ -1,0 +1,3 @@
+import type { CheckboxInputProps } from '../types';
+
+export type ToggleProps = CheckboxInputProps;
