+++ src/unknown-content/index.ts
@@ -1,0 +1,2 @@
+export type { UnknownContentProps } from './types';
+export { UnknownContent } from './unknown-content';
