+++ src/unknown-content/types.ts
@@ -1,0 +1,9 @@
+import type { HTMLAttributes, ReactNode } from 'react';
+
+export interface UnknownContentProps extends HTMLAttributes<HTMLDivElement> {
+  /** Строка с html-версткой. */
+  markup?: string;
+
+  /** Содержимое. */
+  children?: ReactNode;
+}
